front: separate stdcm and operationnal studies
- move MODES, DEFAULT_MODE and associated types from operationnalStudies/consts.ts to main/consts.ts
- move OsrdConfState and OsrdStdcmConfState from operationnalStudies/consts.ts to reducers/osrdconf/consts.ts
- move RUNTIME_CAP from operationnalStudies/consts.ts to stdcm/consts.ts
- move allowancesConsts.ts from applications/stdcm to modules/stdcmAllowances
- update imports, add type keyword when needed
