+++ front/public/locales/en/operationalStudies/manageTrainSchedule.json
@@ -1,13 +1,11 @@
 {
   "addVias": "Add waypoints",
   "addTrainSchedule": "Add one or several trains",
-  "apply": "Apply",
   "blocktype": "Signalling block type",
   "BoundsAreLinked": "Both bounds are linked",
   "BoundsAreNotLinked": "Both bounds are not linked",
   "returnToSimulationResults": "Go back to simulation results",
   "cancelPathfinding": "Cancel",
-  "cancelRequest": "Cancel request",
   "chooseOnMap": "Choose on the map",
   "composition": "Composition",
   "deleteVias": "Delete steps",
@@ -33,8 +31,6 @@
     "noTrainStep": "The increment step of the train must be greater than zero",
     "noTrainStepInput": "Must be > 0",
     "softErrorInfra": "The application is unavailable, a server error has occurred, please try again later.",
-    "stdcmError": "Calculation error for last minute train path.",
-    "stdcmErrorNoPaths": "Incompatibility with other train paths.",
     "The requested train couldn't reach its destination": "The train cannot reach its destination",
     "trainScheduleTitle": "An error has occurred",
     "unableToDeleteTrain": "Impossible to delete the train",
@@ -79,19 +75,13 @@
   "pathfindingMissingParamsSimple": "Missing information for the search",
   "pathfindingNoState": "No route search.",
   "pathfindingStatus": "Pathfinding state",
-  "pleaseWait": "Please wait…",
   "powerRestriction": "Power restriction code",
   "powerRestrictionExplanationText": "By default, the simulation will use the nominal power.",
   "powerRestrictionEmptyExplanationText": "The selected rolling stock does not have any power restrictions. You can edit it in the rolling stock editor.",
   "restartPathfinding": "Restart search",
   "rollingstock": "Rolling stock",
   "simulation": "Simulation",
-  "spaceSpeedGraphic": "Space-Velocity graph",
-  "spaceTimeGraphic": "Space-Time graph",
   "speedLimitByTag": "Composition code",
-  "stdcmComputation": "Search for a train path",
-  "stdcmNoResults": "No path found",
-  "stdcmResults": "Results",
   "stopSimulation": "Stop",
   "tabs": {
     "rollingStock": "Rolling stock",
