+++ front/public/locales/fr/operationalStudies/manageTrainSchedule.json
@@ -1,13 +1,10 @@
 {
   "addVias": "Ajout de vias",
   "addTrainSchedule": "Ajouter un ou plusieurs trains",
-  "apply": "Appliquer",
   "blocktype": "Type de block",
   "BoundsAreLinked": "Les deux bornes sont liées",
   "BoundsAreNotLinked": "Les deux bornes ne sont pas liées",
   "returnToSimulationResults": "Retourner aux résultats de simulation",
-  "cancelPathfinding": "Annuler",
-  "cancelRequest": "Annuler la requête",
   "chooseOnMap": "Choisir sur la carte",
   "composition": "Composition",
   "deleteRoute": "Supprimer l'itinéraire",
@@ -33,8 +30,6 @@
     "noTrainStep": "Le pas d'incrément du nom du train doit être supérieur à zéro",
     "noTrainStepInput": "Doit être > 0",
     "softErrorInfra": "L'application est indisponible, une erreur serveur est survenue, veuillez réessayer ultérieurement.",
-    "stdcmError": "Erreur de calcul Sillon de dernière minute",
-    "stdcmErrorNoPaths": "Incompatibilité avec d'autres sillons.",
     "The requested train couldn't reach its destination": "Le train ne peut pas atteindre sa destination",
     "trainScheduleTitle": "Une erreur est survenue",
     "unableToDeleteTrain": "Impossible de supprimer le train",
@@ -86,12 +81,7 @@
   "restartPathfinding": "Relancer la recherche",
   "rollingstock": "Matériel",
   "simulation": "Simulation",
-  "spaceSpeedGraphic": "Graphique Espace-Vitesse",
-  "spaceTimeGraphic": "Graphique Espace-Temps",
   "speedLimitByTag": "Code de composition",
-  "stdcmComputation": "Recherche de sillon",
-  "stdcmNoResults": "Aucun sillon trouvé",
-  "stdcmResults": "Résultats",
   "stopSimulation": "Stopper",
   "tabs": {
     "rollingStock": "Matériel",
