+++ front/src/modules/stdcmAllowances/allowancesConsts.ts
