+++ front/src/reducers/osrdconf/operationalStudiesConf/index.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import type { OsrdConfState } from 'applications/operationalStudies/consts';
+import type { OsrdConfState } from 'reducers/osrdconf/consts';
 
 import { defaultCommonConf, buildCommonConfReducers } from 'reducers/osrdconf/osrdConfCommon';
