+++ front/src/reducers/osrdconf/osrdConfCommon/__tests__/commonConfBuilder.ts
@@ -1,13 +1,9 @@
 import type { SwitchType } from 'types';
 
-import type {
-  OsrdConfState,
-  PointOnMap,
-  PowerRestrictionRange,
-} from 'applications/operationalStudies/consts';
-
+import type { PointOnMap, PowerRestrictionRange } from 'applications/operationalStudies/consts';
 import type { Allowance, PathResponse } from 'common/api/osrdEditoastApi';
 import { Feature } from 'geojson';
+import type { OsrdConfState } from 'reducers/osrdconf/consts';
 
 export default function commonConfBuilder() {
   return {
