+++ front/src/reducers/osrdconf/stdcmConf/stdcmConfReducers.spec.ts
@@ -1,12 +1,11 @@
-import type { OsrdStdcmConfState, StandardAllowance } from 'applications/operationalStudies/consts';
 import { createStoreWithoutMiddleware } from 'store';
-
 import { describe, expect } from 'vitest';
 import {
   stdcmConfInitialState,
   stdcmConfSlice,
   stdcmConfSliceActions,
 } from 'reducers/osrdconf/stdcmConf';
+import type { OsrdStdcmConfState, StandardAllowance } from 'reducers/osrdconf/consts';
 import testCommonConfReducers from '../osrdConfCommon/__tests__/utils';
 
 const createStore = (initialStateExtra?: Partial<OsrdStdcmConfState>) =>
